kucoin: convert orderbook to TS
